Conflicts:
	src/unsw/dungeon/Equipable.java
	src/unsw/dungeon/Player.java

+++ src/unsw/dungeon/MoveSpeed.java
@@ -1,5 +1,5 @@
 package unsw.dungeon;
 
 public interface MoveSpeed {
-	public int getSpeed();
+	public long getSpeed();
 }
